Moved UserBanCommand to the moderation category
